+++ src/types/UserProjectData.ts
@@ -1,0 +1,7 @@
+export interface UserProjectData {
+    org : string,
+    user: string,
+    project_name : string,
+    guidelines : string,
+    resources : string,
+}