+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Polyverse Boost API on Node.js
 ## Version 0.5.1: January 2nd, 2024
 ### New Features
 - Added APIs /user_project/:org/:project goals, analysis and data_references to delete user project data (DELETE)
-- Added placeholder API /user_project/:org/:project/data_references (POST) to create/upload user data
+- Added API /user_project/:org/:project/data (GET/POST) to create/upload user data (e.g. project data, aispec, blueprint, etc)
 
 ### Enhancements
 - Renamed API from get_file_from_uri to user_project_file (GET)